Added a more robust interface for the IdentityService
Added the ability to set the from address and replyTo in different send
methods. Also added a default from and default replyto to the service
options since From as to be defined for the mailgun service to work.
Also, force a From address to be specified in the simple constructor